clk: qcom: gcc-msm8939: Increment earlycon use count of gcc_blsp1_ahb_clk
The early UART console makes use of the gcc_blsp1_ahb_clk clock for the
UART but, earlycon runs well before the gcc driver so there's no good way
to represent the dependency on the clock.

The dependency on the clock becomes important because the BAM DMA driver
has runtime PM and has a reference to the gcc_blsp1_ahb_clk clock also.
What can happen is the BAM DMA PM runtime code runs and switches off the
gcc_blsp1_ahb_clk clock that the earlycon depends on.

earlycon then resets the system as the IP block is no longer clocked.

Sometimes the MSM UART will successfully probe before the BAM DMA PM code
runs and the system doesn't hang which is a racy behavior we can't rely on.

This patch looks at the kernel command line and if it sees the string for
msm_serial adds a clk_prepare_enable() for gcc_blsp1_ahb_clk.

A DT entry is added to the serial port earlycon is being run from so the
msm_serial driver can know to decrement the reference count.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>