tty: msm_serial: Add earlycon_pclk_disable logic
earlycon on platforms such as msm8916 and msm8939 makes use of a shared
clock the BLSP1_AHB_CLK_ENA called "pclk" here.

Since this clock is shared it is possible for a non-UART driver to disable
it prior to the main msm_serial code running. In the msm8939 clock driver
we increment the reference count to this clock to represent the earlycon
dependency.

The reference count should correspondingly be decremented when the main
msm_serial driver takes over. If the dt bool "qcom,earlycon-pclk-disable"
appears in the DT for a given port, the reference count for the pclk will
be decremented exactly once, after the main driver has taken the references
it needs to via clk_prepare_enable();

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>