ASoC: Add AK4375 support
AK4375 is a 32-bit stereo DAC with headphone amplifier.
There's no documentation for it on akm.com, and only a brief
datasheet can be found floating on the internets [1].

Thanks to Oriane BAYERD <obayerd@eurocomposant.fr>
for finally answering my inquiries, if only to tell me that
this chip is EOL following AKM factory burning in october 2020
and thus no detailed documentation is available anymore...

So this driver is solely based on downstream code [2] and [3]
by Hu Jin from AKM (no known email).

[1] https://datasheetspdf.com/pdf-file/1400317/AKM/AK4375A/1
[2] https://github.com/msm8916-mainline/android_kernel_qcom_msm8916/blob/alcatel-idol347/sound/soc/codecs/idol347/ak4375.c
[3] https://github.com/msm8916-mainline/android_kernel_qcom_msm8916/blob/alcatel-idol347/sound/soc/codecs/ak4375.c

Signed-off-by: Vincent Knecht <vincent.knecht@mailoo.org>